update fixtures, skips, and small bug fixes
